examples/example_one/requirements.txt
Django==2.2.17
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
factory-boy==3.1.0
django-filter==2.4.0
django-cors-headers==3.5.0
ipdb==0.13.4
